Separated data collection dags (hourly and daily)
